[SJ] - can load all the basic data

+++ src/main/java/com/nwu/data/taxi/domain/repository/GPSReadingRepository.java
@@ -3,5 +3,5 @@ package com.nwu.data.taxi.domain.repository;
 import com.nwu.data.taxi.domain.model.GPSReading;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface GPSReadingRepository extends PagingAndSortingRepository<GPSReading, Long> {
+public interface GPSReadingRepository extends PagingAndSortingRepository<GPSReading, Integer> {
 }
